Added Project Team Assembly functionality and fixed a few security issues

+++ app/Http/Controllers/ProjectController.php
@@ -3,24 +3,30 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\CreateProjectRequest;
+use App\Http\Requests\OwnershipRequest;
+use App\Http\Requests\UpdateProjectRequest;
 use App\Models\Academy;
+use App\Models\Application;
 use App\Models\Project;
 use App\Models\User;
 use Illuminate\Http\Request;
 
 class ProjectController extends Controller
 {
-    public function index() {
-        $projects = Project::where('created_by', auth()->user()->id)->orderBy('id', 'DESC')->get();
+    public function index()
+    {
+        $projects = Project::latest()->where('created_by', auth()->user()->id)->get();
         return view('projects.index', compact('projects'));
     }
 
-    public function create() {
+    public function create()
+    {
         $academies = Academy::all();
         return view('projects.create', compact('academies'));
     }
 
-    public function store(CreateProjectRequest $request) {
+    public function store(CreateProjectRequest $request)
+    {
         $project = Project::create([
             'name' => $request->name,
             'description' => $request->description,
@@ -31,31 +37,76 @@ class ProjectController extends Controller
         return redirect()->route('projects');
     }
 
-    public function edit($id) {
+    public function edit($id)
+    {
         $project = Project::find($id);
-        $academies = Academy::all();
-        return view('projects.edit', compact('project', 'academies'));
+        if ($project->created_by == auth()->user()->id) {
+            $academies = Academy::all();
+            return view('projects.edit', compact('project', 'academies'));
+        } else {
+            abort(403, 'Unauthorized');
+        }
     }
 
-    public function update(CreateProjectRequest $request, $id) {
+    public function update(CreateProjectRequest $request, $id)
+    {
         $project = Project::find($id);
-        $project->update([
-            'name' => $request->name,
-            'description' => $request->description,
-        ]);
+        if ($project->created_by == auth()->user()->id) {
+            $project->update([
+                'name' => $request->name,
+                'description' => $request->description,
+            ]);
 
-        $project->academies()->sync($request->academies);
+            $project->academies()->sync($request->academies);
 
-        return redirect()->route('projects');
+            return redirect()->route('projects');
+        } else {
+            abort(403, 'Unauthorized');
+        }
     }
 
-    public function destroy($id) {
+    public function destroy($id)
+    {
+        $project = Project::find($id);
+        if ($project->created_by == auth()->user()->id) {
+            $project->academies()->detach();
+
+            Application::where('project_id', $id)->each(function ($product, $key) {
+                $product->delete();
+            });
+
+            $project->delete();
+
+            return redirect()->route('projects');
+        } else {
+            abort(403, 'Unauthorized');
+        }
+    }
+
+    public function applicants($id)
+    {
+        $project = Project::find($id);
+        if ($project->created_by == auth()->user()->id) {
+            $applications = Application::where('project_id', $id)->get();
+
+            return view('projects.applicants', compact('applications', 'project'));
+        } else {
+            abort(403, 'Unauthorized');
+        }
+    }
+
+    public function assemble($id) {
         $project = Project::find($id);
 
-        $project->academies()->detach();
+        $project->is_assembled = 1;
 
-        $project->delete();
+        $project->save();
 
-        return redirect()->route('projects');
+        Application::where('project_id', $id)->where('is_accepted', null)->each(function ($product, $key) {
+            $product->is_accepted = 0;
+            $product->save();
+        });
+
+        return redirect()->route('projects')->with(['message' => 'Project successfully assembled!']);
     }
 }
